Fuck me

+++ plugins/CompleteTimestamps.plugin.js
@@ -30,7 +30,7 @@ class CompleteTimestamps {
 
 	getDescription () {return "Replace all timestamps with complete timestamps.";}
 
-	getVersion () {return "1.2.7";}
+	getVersion () {return "1.2.8";}
 
 	getAuthor () {return "DevilBro";}
 	
@@ -116,8 +116,8 @@ class CompleteTimestamps {
 						if (change.removedNodes) {
 							change.removedNodes.forEach((node) => {
 								if (this.updateTimestamps && node.tagName && node.getAttribute("layer-id") == "user-settings") {
-									this.setMaxWidth();
 									document.querySelectorAll(".complete-timestamp").forEach(timestamp => {timestamp.classList.remove("complete-timestamp");});
+									document.querySelectorAll(".complete-timestamp-divider").forEach(divider => {divider.remove();});
 									document.querySelectorAll("time").forEach(stamp => {this.changeTimestamp(stamp);})
 									this.updateTimestamps = false;
 								}
@@ -132,8 +132,6 @@ class CompleteTimestamps {
 				{"own":	{name:"Own",		id:"own",		integrated:false,	dic:false}},
 				BDFDB.languages
 			);
-			
-			this.setMaxWidth();
 			
 			$(document)
 				.on("mouseenter." + this.getName(), BDFDB.dotCNS.message + BDFDB.dotCN.messagecontent, (e) => {
@@ -172,6 +170,7 @@ class CompleteTimestamps {
 	stop () {
 		if (typeof BDFDB === "object") {
 			document.querySelectorAll(".complete-timestamp").forEach(stamp => {stamp.classList.remove("complete-timestamp");});
+			document.querySelectorAll(".complete-timestamp-divider").forEach(divider => {divider.remove();});
 			
 			BDFDB.removeLocalStyle(this.getName() + "CompactCorrection");
 			
@@ -262,10 +261,20 @@ class CompleteTimestamps {
 	changeTimestamp (stamp) {
 		if (!stamp.className || stamp.className.toLowerCase().indexOf("timestamp") == -1 || stamp.classList.contains("complete-timestamp")) return;
 		let time = stamp.getAttribute("datetime");
-		if (!time) return;
-		let choice = BDFDB.getData("creationDateLang", this, "choices");
-		stamp.classList.add("complete-timestamp");
-		BDFDB.setInnerText(stamp, this.getTimestamp(this.languages[choice].id, time));
+		if (time) {
+			this.setMaxWidth();
+			let choice = BDFDB.getData("creationDateLang", this, "choices");
+			stamp.classList.add("complete-timestamp");
+			let stampdivider = document.createElement("span");
+			stampdivider.className = "complete-timestamp-divider arabic-fix";
+			stampdivider.style.setProperty("display", "inline", "important");
+			stampdivider.style.setProperty("height", "0px", "important");
+			stampdivider.style.setProperty("width", "0px", "important");
+			stampdivider.style.setProperty("font-size", "0px", "important");
+			stampdivider.innerText = "ARABIC FIX";
+			stamp.parentElement.insertBefore(stampdivider, stamp);
+			BDFDB.setInnerText(stamp, this.getTimestamp(this.languages[choice].id, time));
+		}
 	}
 	
 	getMessageGroup (div) {
@@ -340,31 +349,34 @@ class CompleteTimestamps {
 	}
 	
 	setMaxWidth () {
-		var timestamp = document.querySelector(BDFDB.dotCN.messagetimestampcompact);
-		if (timestamp) {
-			var choice = BDFDB.getData("creationDateLang", this, "choices");
-			var testtimestamp = $(`<time class="${timestamp.className}" style="width: auto !important;">${this.getTimestamp(this.languages[choice].id, new Date(253402124399995))}</time>`);
-			$(testtimestamp).appendTo(document.body);
-			var width = testtimestamp.outerWidth() + 5;
-			testtimestamp.remove();
-			BDFDB.appendLocalStyle(this.getName() + "CompactCorrection", `
-				${BDFDB.dotCN.messagetimestampcompact} {
-					width: ${width}px !important;
-				}
-				${BDFDB.dotCN.messagetimestampcompactismentioned} {
-					width: ${width + 2}px !important;
-				}
-				${BDFDB.dotCN.messagemarkupiscompact} {
-					margin-left: ${width}px !important;
-					text-indent: -${width}px !important;
-				}
-				${BDFDB.dotCN.messageaccessorycompact} {
-					padding-left: ${width}px !important;
-				}
-			`);
-		}
-		else {
-			BDFDB.removeLocalStyle(this.getName() + "CompactCorrection");
+		if (this.currentMode != BDFDB.getDiscordMode()) {
+			this.currentMode = BDFDB.getDiscordMode();
+			var timestamp = document.querySelector(BDFDB.dotCN.messagetimestampcompact);
+			if (timestamp) {
+				var choice = BDFDB.getData("creationDateLang", this, "choices");
+				var testtimestamp = $(`<time class="${timestamp.className}" style="width: auto !important;">${this.getTimestamp(this.languages[choice].id, new Date(253402124399995))}</time>`);
+				$(testtimestamp).appendTo(document.body);
+				var width = testtimestamp.outerWidth() + 5;
+				testtimestamp.remove();
+				BDFDB.appendLocalStyle(this.getName() + "CompactCorrection", `
+					${BDFDB.dotCN.messagetimestampcompact} {
+						width: ${width}px !important;
+					}
+					${BDFDB.dotCN.messagetimestampcompactismentioned} {
+						width: ${width + 2}px !important;
+					}
+					${BDFDB.dotCN.messagemarkupiscompact} {
+						margin-left: ${width}px !important;
+						text-indent: -${width}px !important;
+					}
+					${BDFDB.dotCN.messageaccessorycompact} {
+						padding-left: ${width}px !important;
+					}
+				`);
+			}
+			else {
+				BDFDB.removeLocalStyle(this.getName() + "CompactCorrection");
+			}
 		}
 	}
 }
